Implemented arrow keys & key board to navigate through chapters
I want to add an arrow button to page through the chapters. If this the user presses the left arrow button, user can navigate to the previous chapter. If the user presses on the right arrow button, user can navigate to the next chapter. If we are at the first chapter, and if there is a previous book, user should be redirected to the previous book's last chapter. if we are at last chapter of the book, and if there is a next book, user should be redirected to the next book's first chapter. if we are at the very first book, left arrow should be hidden. and right arrow should be hidden for the last book

** Resolved